Fixed a bug with an Action lambda in MainForm-ApplyToImageBaseOperation.cs
I had forgotten to change the second argument of setPixel in
ApplyToImageBaseOperation

+++ src/main/MainForm-ApplyToImageBaseOperation.cs
@@ -62,7 +62,7 @@ namespace ImageProcessingApplication
 					for(int i = 0; i < width; i++)
 					{
 						int[] aX = array[i];
-						Action<int,byte> setColor = (y,c) => setColorBase(i,y,c);
+						Action<int,int> setColor = (y,c) => setColorBase(i,y,c);
 						for(int j=0; j < height; j++)
 						{
 							setColor(j, aX[j]);
